# docs/swagger.yaml
basePath: /
definitions:
...
  merchant.MerchantFeeResponse:
    properties:
      amount:
        type: number
      amountType:
        type: string
      createdAt:
        type: string
      deductionType:
        type: string
      merchantId:
        type: string
      paymentMethod:
        type: string
      percentage:
        type: number
      reference:
        type: string
      taxPercentage:
        type: number
      taxType:
        type: string
      updatedAt:
        type: string
      uuid:
        type: string
    type: object
  merchant.MerchantRequest:
    properties:
      address:
        maxLength: 254
        type: string
      businessCountry:
        type: string
      businessStructure:
        type: string
      businessType:
        type: string
      description:
        type: string
      districtId:
        minimum: 1
        type: integer
      logo:
        type: string
      merchantEmail:
        type: string
      merchantPhone:
        type: string
      name:
        type: string
      parentID:
        type: string
      picEmail:
        type: string
      picJobTitle:
        type: string
      picName:
        type: string
      picPhone:
        type: string
      postcode:
        type: string
      requesterID:
        type: string
      requesterType:
        type: string
      shortName:
        maxLength: 25
        type: string
      userId:
        type: string
    required:
    - address
    - businessCountry
    - businessStructure
    - businessType
    - districtId
    - logo
    - merchantEmail
    - merchantPhone
    - name
    - picEmail
    - picName
    - picPhone
    - postcode
    - shortName
    type: object
...
  response.Response:
    properties:
      code:
        example: "00"
        type: string
      data: {}
      errors: {}
      meta: {}
    type: object
...
host: localhost
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    To support payment gateway needs, we prepare technical documents on the backend-portal.
    Where the backend-portal is intended to serve services to OpenAPI and MerchantPortal,
    apart from that this service is integrated with Orchestrator, Core Processor
    (SNAP Core Processor and Credit Card Processor)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Portal API
  version: "1.0"
paths:
...
  /crm/v1/merchants/fee/{id}:
    put:
      consumes:
      - application/json
      description: Update merchant fee endpoint
      operationId: merchant-fee-update
      parameters:
      - description: id of merchant fee
        in: path
        name: id
        required: true
        type: string
      - description: JSON Body for Update Merchant fee
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/merchant.MerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/merchant.MerchantFeeResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update merchant fee endpoint
      tags:
      - API - Merchant Fee
...
schemes:
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"